requirement.txt
et-xmlfile==1.1.0
lxml==4.9.3
openpyxl==3.1.2
Pillow==10.0.0
PyPDF2==3.0.1
python-pptx==0.6.21
xlrd==2.0.1
XlsxWriter==3.1.2
